refactor: design-system 폴더 생성
[#64]
