+++ packages/frontend/src/design-system/components/common/index.ts
@@ -1,0 +1,2 @@
+export { default as Button } from "./Button";
+export { Badge, BadgeGroup } from "./Badge";
